Refactoring KEXT_PATCH. Change numerous CHAR8* CHAR16* to XString.

+++ rEFIt_UEFI/cpp_util/memory.cpp
@@ -10,15 +10,16 @@ extern "C" {
  */
 
 #if 1 //!defined(_MSC_VER)
+
 #ifdef __GNUC__
-void* memset(void* dst, int ch, UINT64 count) __attribute__ ((used));
-//void* memcpy(void* dst, const void* src, UINT64 count) __attribute__ ((used));
+  void* memset(void* dst, int ch, size_t count) __attribute__ ((used));
+  //void* memcpy(void* dst, const void* src, UINT64 count) __attribute__ ((used));
 #else
   void* memset(void* dst, int ch, UINT64 count);
-//  void* memcpy(void* dst, const void* src, UINT64 count);
+  //void* memcpy(void* dst, const void* src, UINT64 count);
 #endif
 
-void* memset(void* dst, int ch, UINT64 count)
+void* memset(void* dst, int ch, size_t count)
 {
   SetMem(dst, count, (UINT8)(ch));
   return dst;
